+++ kraftwerk-core/src/main/java/fr/insee/kraftwerk/core/KraftwerkError.java
@@ -1,0 +1,5 @@
+package fr.insee.kraftwerk.core;
+
+public abstract class KraftwerkError {
+
+}