+++ kraftwerk-core/src/main/java/fr/insee/kraftwerk/core/dataprocessing/DataProcessing.java
@@ -3,8 +3,8 @@ package fr.insee.kraftwerk.core.dataprocessing;
 import java.nio.file.Path;
 import java.util.List;
 
+import fr.insee.kraftwerk.core.KraftwerkError;
 import fr.insee.kraftwerk.core.utils.TextFileReader;
-import fr.insee.kraftwerk.core.vtl.ErrorVtlTransformation;
 import fr.insee.kraftwerk.core.vtl.VtlBindings;
 import fr.insee.kraftwerk.core.vtl.VtlExecute;
 import fr.insee.kraftwerk.core.vtl.VtlScript;
@@ -32,7 +32,7 @@ public abstract class DataProcessing {
 
     public abstract String getStepName();
 
-    public String applyVtlTransformations(String bindingName, Path userVtlInstructionsPath, List<ErrorVtlTransformation> errors){
+    public String applyVtlTransformations(String bindingName, Path userVtlInstructionsPath, List<KraftwerkError> errors){
         // First step
         String automatedVtlInstructions = applyAutomatedVtlInstructions(bindingName, errors);
         // Second step
@@ -56,17 +56,16 @@ public abstract class DataProcessing {
      */
     protected abstract VtlScript generateVtlInstructions(String bindingName);
 
-    protected String applyAutomatedVtlInstructions(String bindingName, List<ErrorVtlTransformation> errors){
+    protected String applyAutomatedVtlInstructions(String bindingName, List<KraftwerkError> errors){
         VtlScript automatedInstructions = generateVtlInstructions(bindingName);
-        log.debug(String.format("Automated VTL instructions generated for step %s: see temp file", getStepName(),
-                automatedInstructions));
+        log.debug(String.format("Automated VTL instructions generated for step %s: see temp file", getStepName()));
         if (!(automatedInstructions.isEmpty() || automatedInstructions.toString().contentEquals(""))) {
         	vtlExecute.evalVtlScript(automatedInstructions, vtlBindings, errors);
         }
         return automatedInstructions.toString();
     }
 
-    protected void applyUserVtlInstructions(Path userVtlInstructionsPath, List<ErrorVtlTransformation> errors){
+    protected void applyUserVtlInstructions(Path userVtlInstructionsPath, List<KraftwerkError> errors){
         String vtlScript = TextFileReader.readFromPath(userVtlInstructionsPath);
         log.info(String.format("User VTL instructions read for step %s:%n%s", getStepName(),
                 vtlScript));