+++ kraftwerk-core/src/main/java/fr/insee/kraftwerk/core/extradata/reportingdata/CSVReportingDataParser.java
@@ -17,7 +17,6 @@ import com.opencsv.CSVParser;
 import com.opencsv.CSVParserBuilder;
 import com.opencsv.CSVReader;
 import com.opencsv.CSVReaderBuilder;
-import com.opencsv.ICSVParser;
 import com.opencsv.exceptions.CsvValidationException;
 
 import fr.insee.kraftwerk.core.exceptions.NullException;
@@ -49,8 +48,10 @@ public class CSVReportingDataParser extends ReportingDataParser {
 						ReportingDataUE reportingDataUE1 = reportingData.getListReportingDataUE().stream().filter(
 								reportingDataUEToSearch -> rowIdentifier.equals(reportingDataUEToSearch.getIdentifier()))
 								.findAny().orElse(null);
-						reportingDataUE1.addState(state);
-						reportingDataUE1.sortStates();
+						if (reportingDataUE1 != null) {
+							reportingDataUE1.addState(state);
+							reportingDataUE1.sortStates();
+						}
 						continue;
 					}
 					ReportingDataUE reportingDataUE = new ReportingDataUE(rowIdentifier);
@@ -59,12 +60,12 @@ public class CSVReportingDataParser extends ReportingDataParser {
 				}
 				integrateReportingDataIntoUE(data, reportingData);
 			} else {
-				log.error(String.format("Following CSV file is malformed : %s", new Object[] { filePath }), "");
+				log.error("Following CSV file is malformed : {}", filePath);
 			}
 		} catch (CsvValidationException e) {
-			log.error(String.format("Following CSV file is malformed : %s", new Object[] { filePath }), (Throwable) e);
+			log.error("Following CSV file is malformed : {}, CsvValidationException {} ", filePath, e.getMessage());
 		} catch (IOException e) {
-			log.error(String.format("Could not connect to data file %s", new Object[] { filePath }), e);
+			log.error("Could not connect to data file {} because IOException {}", filePath, e.getMessage());
 		}
 	}
 
@@ -76,6 +77,10 @@ public class CSVReportingDataParser extends ReportingDataParser {
 			parsedDate = dateFormat.parse(rowTimestamp);
 		} catch (ParseException e1) {
 			log.error("Parsing error : {}", e1.getMessage());
+		}
+		if (parsedDate == null) {
+			log.error("Parsing error : the parsed date is null");
+			return 0L;
 		}
 		return TimeUnit.MILLISECONDS.toSeconds(parsedDate.getTime());
 	}
@@ -92,11 +97,9 @@ public class CSVReportingDataParser extends ReportingDataParser {
 			FileReader filereader = new FileReader(filePath.toString());
 			this.csvReader = new CSVReader(filereader);
 			CSVParser parser = (new CSVParserBuilder()).withSeparator(',').build();
-			this.csvReader = (new CSVReaderBuilder(filereader))
-
-					.withCSVParser((ICSVParser) parser).build();
+			this.csvReader = (new CSVReaderBuilder(filereader)).withCSVParser(parser).build();
 		} catch (FileNotFoundException e) {
-			log.error(String.format("Unable to find the file %s", new Object[] { filePath }), e);
+			log.error("Unable to find the file {}, FileNotFoundException {}", filePath, e);
 		}
 	}
 }