+++ kraftwerk-core/src/main/java/fr/insee/kraftwerk/core/sequence/UnimodalSequence.java
@@ -1,7 +1,9 @@
 package fr.insee.kraftwerk.core.sequence;
 
 import java.util.List;
+import java.util.Map;
 
+import fr.insee.kraftwerk.core.KraftwerkError;
 import fr.insee.kraftwerk.core.dataprocessing.CalculatedProcessing;
 import fr.insee.kraftwerk.core.dataprocessing.DataProcessing;
 import fr.insee.kraftwerk.core.dataprocessing.DataProcessingManager;
@@ -10,13 +12,13 @@ import fr.insee.kraftwerk.core.dataprocessing.UnimodalDataProcessing;
 import fr.insee.kraftwerk.core.inputs.ModeInputs;
 import fr.insee.kraftwerk.core.inputs.UserInputs;
 import fr.insee.kraftwerk.core.metadata.CalculatedVariables;
+import fr.insee.kraftwerk.core.metadata.ErrorVariableLength;
 import fr.insee.kraftwerk.core.metadata.LunaticReader;
-import fr.insee.kraftwerk.core.metadata.MetadataUtils;
+import fr.insee.kraftwerk.core.metadata.Variable;
 import fr.insee.kraftwerk.core.metadata.VariablesMap;
 import fr.insee.kraftwerk.core.parsers.DataFormat;
 import fr.insee.kraftwerk.core.utils.FileUtils;
 import fr.insee.kraftwerk.core.utils.TextFileWriter;
-import fr.insee.kraftwerk.core.vtl.ErrorVtlTransformation;
 import fr.insee.kraftwerk.core.vtl.VtlBindings;
 import lombok.NoArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -26,10 +28,22 @@ import lombok.extern.slf4j.Slf4j;
 public class UnimodalSequence {
 
 	public void unimodalProcessing(UserInputs userInputs, String dataMode, VtlBindings vtlBindings,
-			List<ErrorVtlTransformation> errors) {
+								   List<KraftwerkError> errors, Map<String, VariablesMap> metadataVariables) {
 		ModeInputs modeInputs = userInputs.getModeInputs(dataMode);
-		VariablesMap metadata = MetadataUtils.getMetadata(userInputs, dataMode);
 		String vtlGenerate;
+
+		/* Step 2.4a : Check incoherence between expected variables' length and actual length received */
+		VariablesMap variablesMap = metadataVariables.get(dataMode);
+		for (String variableName : variablesMap.getVariableNames()){
+			Variable variable = variablesMap.getVariable(variableName);
+			if (!(variable.getSasFormat() == null) && variable.getExpectedLength()<variable.getMaxLengthData()){
+				log.warn(String.format("%s expected length is %s but max length received is %d",variable.getName(),variable.getExpectedLength(), variable.getMaxLengthData()));
+				ErrorVariableLength error = new ErrorVariableLength(variable, dataMode);
+				if (!errors.contains(error)){
+					errors.add(error);
+				}
+			}
+		}
 
 		/* Step 2.4b : Apply VTL expression for calculated variables (if any) */
 		if (modeInputs.getLunaticFile() != null) {
@@ -50,12 +64,12 @@ public class UnimodalSequence {
 		}
 
 		/* Step 2.4c : Prefix variable names with their belonging group names */
-		vtlGenerate = new GroupProcessing(vtlBindings, metadata).applyVtlTransformations(dataMode, null, errors);
+		vtlGenerate = new GroupProcessing(vtlBindings, variablesMap).applyVtlTransformations(dataMode, null, errors);
 		TextFileWriter.writeFile(FileUtils.getTempVtlFilePath(userInputs, "GroupProcessing", dataMode), vtlGenerate);
 
 		/* Step 2.5 : Apply mode-specific VTL transformations */
 		UnimodalDataProcessing dataProcessing = DataProcessingManager.getProcessingClass(modeInputs.getDataFormat(),
-				vtlBindings, metadata);
+				vtlBindings, variablesMap);
 		vtlGenerate = dataProcessing.applyVtlTransformations(dataMode, modeInputs.getModeVtlFile(), errors);
 		TextFileWriter.writeFile(FileUtils.getTempVtlFilePath(userInputs, dataProcessing.getStepName(), dataMode),
 				vtlGenerate);