+++ kraftwerk-core/src/test/java/fr/insee/kraftwerk/core/dataprocessing/CalculatedProcessingTest.java
@@ -8,6 +8,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import java.util.ArrayList;
 import java.util.List;
 
+import fr.insee.kraftwerk.core.KraftwerkError;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
@@ -16,7 +17,6 @@ import fr.insee.kraftwerk.core.metadata.CalculatedVariables.CalculatedVariable;
 import fr.insee.kraftwerk.core.metadata.Variable;
 import fr.insee.kraftwerk.core.metadata.VariableType;
 import fr.insee.kraftwerk.core.metadata.VariablesMap;
-import fr.insee.kraftwerk.core.vtl.ErrorVtlTransformation;
 import fr.insee.kraftwerk.core.vtl.VtlBindings;
 import fr.insee.kraftwerk.core.vtl.VtlScript;
 import fr.insee.vtl.model.Dataset;
@@ -29,7 +29,7 @@ public class CalculatedProcessingTest {
     private static CalculatedVariables fooCalculated;
     private static VariablesMap fooVariables;
     private static VtlBindings vtlBindings;
-    private static List<ErrorVtlTransformation> errors;
+    private static List<KraftwerkError> errors;
 
     @BeforeAll
     static void setFooCalculated() {