+++ kraftwerk-core/src/test/java/fr/insee/kraftwerk/core/parsers/PaperDataParserTest.java
@@ -1,7 +1,0 @@
-package fr.insee.kraftwerk.core.parsers;
-
-public class PaperDataParserTest {
-
-	// TODO: unit test samples
-
-}